server/requirements.txt
librosa
pandas
numpy
boto3
pydub
tqdm
python-dotenv
psycopg2-binary
sqlalchemy
pgvector
ffmpeg-python
celery==5.3.4
redis==5.0.1
fastapi
uvicorn
bcrypt
python-jose
